Split `resources/api.py` into package mirroring the `resources/models` hierarchy
