+++ resources/api/__init__.py
@@ -1,0 +1,4 @@
+from .base import all_views
+from .resource import ResourceListViewSet, ResourceViewSet, PurposeViewSet
+from .reservation import ReservationViewSet
+from .unit import UnitViewSet
